removed get methods, moved set methods to Hypergraph
